Handle for issue #2.
Changed to check every time when display over other apps permission is required.
Changed so that the display over other apps permission setting screen is not automatically displayed when the application is launched.
When displaying an overlay, if display over other apps is not allowed, a dialog box will now prompt for permission.